Fixed compile time bugs and warnings exposed by Visual Studio.
git-svn-id: http://yate.null.ro/svn/yate/trunk@6256 acf43c95-373e-0410-b603-e72c3f656dc1

+++ yatengine.h
@@ -1559,7 +1559,7 @@ public:
      * @return Average age of dequeued messages
      */
     unsigned int messageAge(bool usec = false) const
-	{ return m_dispatcher.messageAge(usec); }
+	{ return (unsigned int)m_dispatcher.messageAge(usec); }
 
     /**
      * Retrieve dispatcher's statistics counters
